Move label to separated file for clarity
